整个代码优化 (vibe-kanban a999a804)
#### 第一阶段：紧急改进（Month 1-3）🔴 P0

**目标**：解决阻塞性问题，提升稳定性

| 任务 | 工作量 | 预期成果 |
|------|--------|---------|
| **修复 827 unwrap/expect** | 2-3 周 | 零 panic 风险 |
| **减少 1,415 clone** | 1-2 周 | 性能提升 30-50% |
| **清理 1,244 warnings** | 1 周 | warnings < 100 |
| **创建简化 API** | 4 周 | 3 行代码上手 |
| **LangChain 官方集成** | 4 周 | 无缝集成 | 分析有些功能是不是已经实现，实现就跳过

+++ crates/agent-mem-intelligence/src/processing/consolidation.rs
@@ -256,7 +256,7 @@ impl MemoryConsolidator {
 
         // Update the base memory
         memories[base_idx].content = agent_mem_traits::Content::Text(merged_content);
-        memories[base_idx].set_score(merged_importance as f64);
+        memories[base_idx].set_score(merged_importance);
         memories[base_idx].metadata.access_count = merged_access_count;
         memories[base_idx].metadata.updated_at = chrono::Utc::now();
